Server/requirements.txt
django
requests
beautifulsoup4
djangorestframework
django-cors-headers
gunicorn
whitenoise
